@@ -1,6 +1,5 @@
 import React, { ComponentType } from "react";
-import { Switch, Route } from "react-router-dom";
-import { hot } from "react-hot-loader/root";
+import { useRoutes } from "react-router-dom";
 import { HeaderContainer } from "./Containers/HeaderContainer";
 import Footer from "./Components/Footer/Footer";
 import TriggerEditContainer from "./Containers/TriggerEditContainer";
@@ -35,71 +34,121 @@ import styles from "./desktop.less";
 const cn = classNames.bind(styles);
 
 type ResponsiveRouteProps = {
-    exact?: boolean;
-    path: string;
     container: ComponentType<TriggerListProps> | ComponentType<TriggerProps>;
     view: ComponentType<TriggerListDesktopProps> | ComponentType<TriggerDesktopProps>;
 };
 
-function ResponsiveRoute({ container: Container, view: View, ...rest }: ResponsiveRouteProps) {
+function ResponsiveRoute({ container: Container, view: View }: ResponsiveRouteProps) {
     // @ts-ignore problem with typing view
-    return <Route {...rest} render={(props) => <Container {...props} view={View} />} />;
+    return <Container view={View} />;
 }
 
 function Desktop() {
     const { isChristmasMood } = useAppSelector(UIState);
 
+    const routes = useRoutes([
+        {
+            path: getPagePath("index"),
+            element: <ResponsiveRoute container={TriggerList} view={TriggerListDesktop} />,
+        },
+        {
+            path: getPagePath("triggerAdd"),
+            element: <TriggerAddContainer />,
+        },
+        {
+            path: getPagePath("triggerDuplicate"),
+            element: <TriggerDuplicateContainer />,
+        },
+        {
+            path: getPagePath("triggerEdit"),
+            element: <TriggerEditContainer />,
+        },
+        {
+            path: getPagePath("trigger"),
+            element: <ResponsiveRoute container={Trigger} view={TriggerDesktop} />,
+        },
+        {
+            path: getPagePath("teamSettings"),
+            element: <TeamSettingsPrivateRoute Component={SettingsContainer} />,
+        },
+        {
+            path: getPagePath("settings"),
+            element: <SettingsContainer />,
+        },
+        {
+            path: getPagePath("teams"),
+            element: <TeamsContainer />,
+        },
+        {
+            path: getPagePath("allTeams"),
+            element: (
+                <AdminRoute>
+                    <AllTeamsContainer />
+                </AdminRoute>
+            ),
+        },
+        {
+            path: getPagePath("notifications"),
+            element: (
+                <AdminRoute>
+                    <NotificationListContainer />
+                </AdminRoute>
+            ),
+        },
+        {
+            path: getPagePath("tags"),
+            element: (
+                <AdminRoute>
+                    <TagListContainer />
+                </AdminRoute>
+            ),
+        },
+        {
+            path: getPagePath("patterns"),
+            element: (
+                <AdminRoute>
+                    <PatternListContainer />
+                </AdminRoute>
+            ),
+        },
+        {
+            path: getPagePath("contacts"),
+            element: (
+                <AdminRoute>
+                    <ContactsContainer />
+                </AdminRoute>
+            ),
+        },
+        {
+            path: getPagePath("noisiness"),
+            element: (
+                <AdminRoute>
+                    <NoisinessContainer />
+                </AdminRoute>
+            ),
+        },
+        {
+            path: getPagePath("systemSubscriptions"),
+            element: (
+                <AdminRoute>
+                    <SystemSubscriptionsContainer />
+                </AdminRoute>
+            ),
+        },
+        {
+            path: "*",
+            element: <ErrorContainer />,
+        },
+    ]);
+
     return (
         <div className={cn("layout")}>
             <HeaderContainer className={cn("header")} />
             {isChristmasMood && <ChristmasLights />}
-            <Switch>
-                <ResponsiveRoute
-                    exact
-                    path={getPagePath("index")}
-                    container={TriggerList}
-                    view={TriggerListDesktop}
-                />
-                <Route exact path={getPagePath("triggerAdd")} component={TriggerAddContainer} />
-                <Route
-                    exact
-                    path={getPagePath("triggerDuplicate")}
-                    component={TriggerDuplicateContainer}
-                />
-                <Route exact path={getPagePath("triggerEdit")} component={TriggerEditContainer} />
-                <ResponsiveRoute
-                    exact
-                    path={getPagePath("trigger")}
-                    container={Trigger}
-                    view={TriggerDesktop}
-                />
-                <TeamSettingsPrivateRoute
-                    exact
-                    path={getPagePath("teamSettings")}
-                    component={SettingsContainer}
-                />
-                <Route exact path={getPagePath("settings")} component={SettingsContainer} />
-                <Route exact path={getPagePath("teams")} component={TeamsContainer} />
-                <AdminRoute exact path={getPagePath("allTeams")} component={AllTeamsContainer} />
-                <AdminRoute
-                    exact
-                    path={getPagePath("notifications")}
-                    component={NotificationListContainer}
-                />
-                <AdminRoute exact path={getPagePath("tags")} component={TagListContainer} />
-                <AdminRoute exact path={getPagePath("patterns")} component={PatternListContainer} />
-                <AdminRoute exact path={getPagePath("contacts")} component={ContactsContainer} />
-                <AdminRoute exact path={getPagePath("noisiness")} component={NoisinessContainer} />
-                <AdminRoute
-                    exact
-                    path={getPagePath("systemSubscriptions")}
-                    component={SystemSubscriptionsContainer}
-                />
-                <Route component={ErrorContainer} />
-            </Switch>
+            {routes}
             <Footer className={cn("footer")} />
         </div>
     );
 }
 
-export default hot(Desktop);
+export default Desktop;
